+++ frontend/app-example/(tabs)/index.tsx
