+++ frontend/app-example/info.md
@@ -1,0 +1,5 @@
+ # app
+ 
+Contains the app's navigation, which is file-based. The file structure of the app directory determines the app's navigation.
+
+The app has two routes defined by two files: app/(tabs)/index.tsx and app/(tabs)/explore.tsx. The layout file in app/(tabs)/_layout.tsx sets up the tab navigator.
