+++ frontend/assets/info.md
@@ -1,0 +1,2 @@
+# assets
+The assets folder contains adaptive-icon.png used for Android and an icon.png used for iOS as app icons. It also contains splash.png which is an image for the project's splash screen and a favicon.png if the app runs in a browser.
