+++ frontend/components/info.md
@@ -1,0 +1,3 @@
+# components
+
+Contains React Native components, like ThemedText.tsx, which creates text elements that use the app's color scheme in light and dark modes.
