+++ frontend/constants/info.md
@@ -1,0 +1,3 @@
+# constants
+
+Contains Colors.ts, which contains a list of color values throughout the app.
