+++ frontend/hooks/info.md
@@ -1,0 +1,3 @@
+# hooks
+
+Contains React Hooks, which allows sharing common behavior between components. For example, useThemeColor() is a hook that returns a color based on the current theme.
