+++ frontend/scripts/info.md
@@ -1,0 +1,3 @@
+# scripts
+
+Contains reset-project.js, which can be run with npm run reset-project. This script will move the app directory to app-example, and create a new app directory with an index.tsx file.
